gen: move stuff
